refactor: removing all the mutexes from inside the `AppState` struct (#83)
* Removing all the mutexes from inside the AppState struct

The mutex is now moved up one level. So nowhere in the code there needs
to be locking of mutexes anymore.

The unaltered_state is temporarily uncommented. But I doubt this have
ever worked because the whole code was behind mutexes so I guess it
would only clone the pointer the the referenced data. But we maybe need
to tackle this in another PR.

* Fix code to save messages and unit tests

Take a look at the unaltered state changes