feat: use the `redirect_url` and `credential_issuer_url` host to create the `connection_url`
Using the same method to create the `connection_url` for both the `Relying Party`
and the `Credential Issuer` connections ensures that the `connection_url` is
consistent when credentials are issued and requested by the same third-party.

This also means that the `Connection` ID's can be generated now based on both
the issuer/client name AND the `connection_url` (instead of just the `name`).

This solves the problem that different Servers with the same `name` would result
in the same `Connection` ID's.